// autogpt_platform/frontend/src/app/api/__generated__/models/blockListResponse.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { ResponseType } from "./responseType";
import type { BlockListResponseSessionId } from "./blockListResponseSessionId";
import type { BlockInfoSummary } from "./blockInfoSummary";

/**
 * Response for find_block tool.
 */
export interface BlockListResponse {
  type?: ResponseType;
  message: string;
  session_id?: BlockListResponseSessionId;
  blocks: BlockInfoSummary[];
  count: number;
  query: string;
  usage_hint?: string;
}